+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 2.1.2 (...)
 - fix IPv6 address splitting for >16 OpenVPN processes per profile (issue #43)
+- even when `defaultGateway` is `true` push the routes as configured in `route`
+  (issue #44)
 
 ## 2.1.1 (2019-12-10)
 - update for server API to handle per profile tls-crypt keys